+++ src/main/java/com/bookspot/stock/application/LibraryStockService.java
@@ -1,4 +1,4 @@
-package com.bookspot.stock;
+package com.bookspot.stock.application;
 
 import com.bookspot.stock.domain.LibraryStockRepository;
 import lombok.RequiredArgsConstructor;
